Add implicit group pubkey to group signature
A group message doesn't explicitly contain the group that the message
was written for, which allows cross-posting of messages.  This appends
the group pubkey to the signature and verification to eliminate this
possibility.

(It is worth noting that this same implicit recipient pubkey is already
present in 1-1 message signatures, and that the update here conceptually
matches the solution that has been long used in 1-1 messages to prevent
any similar replay of DMs).

The idea here is that if A and B are each in groups G and H, and A
sends a message M to group G then B could take M, decrypt it, leave
the signature intact, and then reencrypt it for group H and send it to
group H.  Members of group H would then see a message that is accepted
as coming from A, even though A did not send such a message to H.

This commit solves it by reworking our group message encryption to
include the implicit pubkey in the signature.  That is, we were
previously signing/verifying just `plaintext` (the blob of data
containing the encoded message data), but now we sign/verify `plaintext
|| group_ed25519_pubkey`.

It is "implicit" in that we don't include that group pubkey as part of
the message, because any legitimate recipient of the message already
knows the group it was received it, and so if they receive it in a
different group the value they verify `plaintext || different_pubkey`
isn't the value that was signed, `plaintext || original_pubkey`, and so
will fail verification of any clients that receive it in the wrong
group.